Don't create one QNetworkAccessManager per forecast query job
Additionally, allow to provide an external QNAM, to reuse one the using
application might already have. If we use the internal one, enable HSTS
on it as well.

This only covers the forecast jobs, but would also need to be propagated
to its subjobs, but the timezone and sunrise ones we can hopefully replace
entirely by offline lookup in KI18n and the astronomical calculation code
in KHolidays eventually.