Make LocationQuery return a job for each request
This makes it possible to properly associate results and errors with a
request, as well as cancel ongoing ones. As a bonus this now also has
error reporting for looking up GPS positions, which makes the unit test
significantly faster when GPS is not available.

This changes ABI, but it API and behavior compatible to not require
changes in KWeather immediately.