Add a proper coordinate type for CAP polygons, rather than using std::pair
"first" and "second" are very error prone when using this, as it's not
immediately clear whether that's x/y or latitude/longitude and you thus
run the risk of swapping them.

While at it, unify the name of the polygon typedef and replace a few
struct uses that trigger warnings on Android due to inconsistent use.