Remove the unused alert feed handling
As per discussion in #8.

This has technical limitations but also conceptual ones, as high-frequency
polling from all clients to those feeds is inefficient and not something
that is meant to be done from the server perspective.

The proper solution for this is fetching CAP feeds server-side and using
push notifications.