fix: Add virtual destructor for ReplyPrivate
This fixes a memory error detectable by valgrind where deletion of
PendingWeatherForecast instances would trigger a deallocation of the
wrong size (not matching the corresponding allocation). This error also
triggers an allocator error and SIGABRT when using hardened_malloc (from
GrapheneOS) as the memory allocator.

The error was because, when a PendingWeatherForecast instance was
deleted, its internal PendingWeatherForecastPrivate would be deallocated
as if it were an instance of its base class ReplyPrivate (due to the
missing virtual destructor).

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>